feat: hoist store & move workflowEditor to new react mount (#3063)
* feat: moving workflow editor & hoisting store

* fix: subscribe only when onChange provided

* feat: get store updates from phoenix

* feat: move store sync into workflowstoresync component

* feat: pushing patch updates to phoenix

* refactor: remove old phoenix-hook WorkflowEditor

* feat: hide form when when in jobEditor view

* fix: pass react-mount container element into context

* refactor: move store to its own directory

* fix: update location of store

* feat: calculate canvas width & height from ReactFlow element

* refactor: remove log
